+++ packages/brick_graphql/lib/src/graphql_adapter.dart
@@ -2,27 +2,11 @@ import 'package:brick_core/core.dart';
 import 'package:brick_graphql/src/graphql_model.dart';
 import 'package:brick_graphql/src/graphql_provider.dart';
 import 'package:brick_graphql/src/runtime_graphql_definition.dart';
-import 'package:gql/ast.dart';
+import 'package:brick_graphql/src/transformers/graphql_query_operation_transformer.dart';
 
 /// Constructors that convert app models to and from REST
 abstract class GraphqlAdapter<_Model extends Model> implements Adapter<_Model> {
-  /// The mutation used to remove data.
-  DocumentNode? get defaultDeleteOperation => null;
-
-  /// The query used to fetch multiple members.
-  DocumentNode? get defaultQueryOperation => null;
-
-  /// The query used to fetch a member.
-  DocumentNode? get defaultQueryFilteredOperation => null;
-
-  /// The subscription used to fetch all members.
-  DocumentNode? get defaultSubscriptionOperation => null;
-
-  /// The subscription used to fetch a member or specific members.
-  DocumentNode? get defaultSubscriptionFilteredOperation => null;
-
-  /// The mutation used to create or update a member.
-  DocumentNode? get defaultUpsertOperation => null;
+  GraphqlQueryOperationTransformer Function(Query?, GraphqlModel?)? get queryOperationTransformer;
 
   Map<String, RuntimeGraphqlDefinition> get fieldsToGraphqlRuntimeDefinition;