+++ packages/brick_graphql/CHANGELOG.md
@@ -1,5 +1,62 @@
 ## Unreleased
 
+## 2.0.0
+
+* **BREAKING CHANGE** `GraphqlProvider#queryToVariables` has been moved to internal class `GraphqlRequest`
+
+**BREAKING CHANGE**
+
+All Graphql operations are now declared in a single class - `GraphqlQueryOperationTransformer`. A single  Migration guide:
+
+1. Create a new class that extends `GraphqlQueryOperationTransformer`:
+    ```dart
+    class UserQueryOperationTransformer extends GraphqlQueryOperationTransformer {
+      const UserQueryOperationTransformer(super.query, super.instance);
+    }
+    ```
+1. This class has access to every request's `query`, and for `delete` and `upsert`, `instance`. Move all declared properties to within one of `get`, `delete`, `subscribe` or `upsert`. `defaultSubscriptionFilteredOperation` and `defaultSubscriptionOperation` are now `subscribe`. Additionally, `defaultQueryFilteredOperation` and `defaultQueryFilteredOperation` have been consolidated to `get` (example below)
+    ```dart
+    class UserQueryOperationTransformer extends GraphqlQueryOperationTransformer {
+      GraphqlOperation get get {
+        if (query.where != null) {
+          return GraphqlOperation(document: r'''
+            query FilteredUsers($name: String!) {
+              usersByName(input: $input) {}
+            }
+          ''');
+        }
+        return GraphqlOperation(document: r'''
+          query AllUsers {
+            users {}
+          }
+        ''');
+      }
+    }
+    ```
+1. In complex cases where the entire model is not being transmitted, `variables` can also be supplied.
+    ```dart
+    class UserQueryOperationTransformer extends GraphqlQueryOperationTransformer {
+      GraphqlOperation get upsert {
+        if (query.where != null) {
+          return GraphqlOperation(
+            document: r'''
+              mutation UpdateUserName($name: String!) {
+                updateUserName(input: $input) {}
+              }
+            ''',
+            variables: {'name': Where.firstByField('name', query.where)});
+        }
+        return null;
+      }
+    }
+    ```
+1. Use the class in `GraphqlSerializable` (this replaces all operation declarations):
+    ```dart
+    @GraphqlSerializable(
+      queryOperationTransformer: UserQueryOperationTransformer.new
+    )
+    ```
+
 ## 1.3.1
 
 * Only return `get` documents for `get` operations when constructing the GraphQL document